refactor: folders into infra
