+++ src/main/java/com/umc/linkyou/awss3/AwsS3Service.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.awsS3;
+package com.umc.linkyou.awss3;
 
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.services.s3.AmazonS3;
@@ -21,7 +21,6 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URLDecoder;
-import java.util.Optional;
 import java.util.UUID;
 
 @Slf4j
