+++ src/main/java/com/umc/linkyou/infra/ai/summary/OpenAISummaryUtil.java
@@ -1,10 +1,11 @@
-package com.umc.linkyou.openApiSummary;
+package com.umc.linkyou.infra.ai.summary;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.umc.linkyou.apiPayload.code.status.ErrorStatus;
 import com.umc.linkyou.apiPayload.exception.GeneralException;
-import com.umc.linkyou.utils.extractors.WebContentExtractor;
+import com.umc.linkyou.infra.ai.dto.SummaryAnalysisResultDTO;
+import com.umc.linkyou.infra.parser.WebContentExtractor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
