+++ src/main/java/com/umc/linkyou/infra/parser/TitleDomainParser.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.utils.extractors;
+package com.umc.linkyou.infra.parser;
 
 import lombok.extern.slf4j.Slf4j;
 import org.jsoup.Jsoup;
