+++ src/main/java/com/umc/linkyou/infra/parser/WebContentExtractor.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.utils.extractors;
+package com.umc.linkyou.infra.parser;
 
 import com.umc.linkyou.domain.classification.Domain;
 import com.umc.linkyou.domain.enums.CrawlStrategy;
