+++ src/main/java/com/umc/linkyou/service/AiArticleServiceImpl.java
@@ -8,9 +8,9 @@ import com.umc.linkyou.domain.classification.*;
 import com.umc.linkyou.domain.Linku;
 import com.umc.linkyou.domain.mapping.SituationJob;
 import com.umc.linkyou.domain.mapping.UsersLinku;
-import com.umc.linkyou.titleImgParser.LinkToImageService;
-import com.umc.linkyou.openApiSummary.OpenAISummaryUtil;
-import com.umc.linkyou.openApiSummary.SummaryAnalysisResultDTO;
+import com.umc.linkyou.infra.parser.LinkToImageService;
+import com.umc.linkyou.infra.ai.summary.OpenAISummaryUtil;
+import com.umc.linkyou.infra.ai.dto.SummaryAnalysisResultDTO;
 import com.umc.linkyou.repository.*;
 import com.umc.linkyou.repository.aiArticleRepository.AiArticleRepository;
 import com.umc.linkyou.repository.linkuRepository.LinkuRepository;
